Added function to extract and process uncorrected images (not just LEDAPS)

+++ extract_and_process.py
@@ -4,11 +4,39 @@ import fnmatch
 from glob import glob
 from subprocess import call
 
-from convert_DN_to_radiance import create_radiance_images
+from convert_DN_to_radiance import create_radiance_image
 from mask import mask_all_bands
 
 import logging
 
+def extract_and_process_uncorrected(uncorrected_fname, path):
+	logging.basicConfig(format='%(asctime)s:%(levelname)s:%(message)s', datefmt='%H:%M:%S', level=logging.DEBUG)
+	# ledaps_fname = r"E:\_Datastore\LandsatCDR\_Originals\LE71790772000097-SC20130626081016.tar.gz"
+	# uncorrected_fname = r"E:\_Datastore\LandsatCDR\_Originals\LE71790772000097EDC01.tar.gz"
+	# path = "E:\_Datastore\LandsatCDR\TestAutomatic"
+
+
+	logging.info("Processing image with scene ID = %s to %s", os.path.basename(uncorrected_fname).replace(".tar.gz", ""), path)
+	# Keep track of what the current working directory is, and change to the path where we want to put the files
+	prev_working_dir = os.getcwd()
+	if not os.path.exists(path):
+		os.mkdir(path)
+	os.chdir(path)
+
+	# Extract the uncorrected .tar.gz file
+	tar = tarfile.open(uncorrected_fname)
+	logging.info("Extracting %s", uncorrected_fname)
+	tar.extractall(path)
+	logging.info("Extraction complete.")
+
+
+	rootname = os.path.basename(uncorrected_fname).replace(".tar.gz", "")
+
+	logging.info("Merging uncorrected images")
+	# Create a merged radiance image from the raw DN tif files
+	create_radiance_image(os.path.join(path, rootname), os.path.join(path,"Uncorrected_Merged.tif"))
+
+	os.chdir(prev_working_dir)
 
 def extract_and_process(ledaps_fname, uncorrected_fname, path):
 	logging.basicConfig(format='%(asctime)s:%(levelname)s:%(message)s', datefmt='%H:%M:%S', level=logging.DEBUG)
@@ -38,7 +66,7 @@ def extract_and_process(ledaps_fname, uncorrected_fname, path):
 
 	logging.info("Merging uncorrected images")
 	# Create a merged radiance image from the raw DN tif files
-	create_radiance_images(os.path.join(path, rootname), os.path.join(path,"Uncorrected_Merged.tif"))
+	create_radiance_image(os.path.join(path, rootname), os.path.join(path,"Uncorrected_Merged.tif"))
 
 	logging.info("Extracting masks from LEDAPS file")
 	# Extract the masks from the LEDAPS lndsr*.hdf file
@@ -78,6 +106,6 @@ def extract_and_process(ledaps_fname, uncorrected_fname, path):
 	logging.info("Removing temporary files")
 	all_files = glob(os.path.join(path, "*.*"))
 	for f in all_files:
-		if not "Final" in f and not "Water" in f and not "Cloud" in f:
+		if not "Final" in f and not "Water" in f and not "Cloud" in f and not "MTL" in f:
 			logging.debug("Removing %s", f)
 			os.remove(f)
